Extract into component.
