+++ src/components/cluster/detail/V4ClusterDetailTable.js
@@ -1,9 +1,9 @@
 import { Code, Dot, FlexRowWithTwoBlocksOnEdges } from 'styles';
 import { CopyToClipboardDiv } from './V5ClusterDetailTable';
 import { getCpusTotal, getMemoryTotal } from 'utils/cluster_utils';
-import AWSAccountID from 'UI/aws_account_id';
 import Button from 'UI/button';
 import copy from 'copy-to-clipboard';
+import CredentialInfoRow from './CredentialInfoRow';
 import moment from 'moment';
 import OverlayTrigger from 'react-bootstrap/lib/OverlayTrigger';
 import PortMappingsRow from './PortMappingsRow';
@@ -105,59 +105,6 @@ class V4ClusterDetailTable extends React.Component {
       api_endpoint,
     } = cluster;
 
-    // TODO Refactor credentials: markup in render, data in variables or functions outside render
-    // BYOC provider credential info
-    let credentialInfoRows = [];
-    if (
-      cluster &&
-      credential_id &&
-      credential_id != '' &&
-      credentials.items.length === 1
-    ) {
-      // check if we have the right credential info
-      if (credentials.items[0].id !== credential_id) {
-        credentialInfoRows.push(
-          <div key='providerCredentialsInvalid'>
-            <div>Provider credentials</div>
-            <div className='value'>
-              Error: cluster credentials do not match organization credentials.
-              Please contact support for details.
-            </div>
-          </div>
-        );
-      } else {
-        if (provider === 'aws') {
-          credentialInfoRows.push(
-            <div key='awsAccountID'>
-              <div>AWS account</div>
-              <div className='value code'>
-                <AWSAccountID
-                  roleARN={credentials.items[0].aws.roles.awsoperator}
-                />
-              </div>
-            </div>
-          );
-        } else if (provider === 'azure') {
-          credentialInfoRows.push(
-            <div key='azureSubscriptionID'>
-              <div>Azure subscription</div>
-              <div className='value code'>
-                {credentials.items[0].azure.credential.subscription_id}
-              </div>
-            </div>
-          );
-          credentialInfoRows.push(
-            <div key='azureTenantID'>
-              <div>Azure tenant</div>
-              <div className='value code'>
-                {credentials.items[0].azure.credential.tenant_id}
-              </div>
-            </div>
-          );
-        }
-      }
-    }
-
     return (
       <WrapperDiv>
         <FlexRowWithTwoBlocksOnEdges>
@@ -227,12 +174,11 @@ class V4ClusterDetailTable extends React.Component {
         </FlexRowWithTwoBlocksOnEdges>
 
         <PortMappingsRow cluster={cluster} />
-
-        {credentialInfoRows.length !== 0 && (
-          <FlexRowWithTwoBlocksOnEdges>
-            {credentialInfoRows}
-          </FlexRowWithTwoBlocksOnEdges>
-        )}
+        <CredentialInfoRow
+          cluster={cluster}
+          credentials={credentials}
+          provider={provider}
+        />
 
         <hr style={{ margin: '25px 0' }} />
         <h2>Worker nodes</h2>
